@@ -1,1 +1,1 @@
-ruby-2.4.4
+ruby-2.6.6
