@@ -1,14 +1,14 @@
 require "rails_helper"
 
-RSpec.describe Api::V1::SlicesController, type: :controller do 
+RSpec.describe Api::V1::SlicesController, type: :controller do
   include_context 'api_schema'
 
   render_views
 
-  let(:widget) { FactoryGirl.create(:widget_with_data, datasets_count: 2) }
+  let(:widget) { FactoryBot.create(:widget_with_data, datasets_count: 2) }
   let(:ds_1) { widget.data_table.datasets.first }
   let(:ds_2) { widget.data_table.datasets.second }
-  let(:user) { FactoryGirl.create(:user_with_token,
+  let(:user) { FactoryBot.create(:user_with_token,
     dashboards: [widget.dashboard])}
 
   let(:authorization) {
@@ -30,15 +30,15 @@ RSpec.describe Api::V1::SlicesController, type: :controller do
     include_examples 'api_authorized_ok'
   end
 
-  describe '#create' do 
+  describe '#create' do
     let(:req) { post :create, params: { widget_id: widget.id, period: 'month',
       period_start: ts,
       groups: [ds_1, ds_2].collect {|ds| { dataset_id: ds.id, value: data_value } } } }
 
-    context 'valid value' do 
+    context 'valid value' do
       let(:data_value) { 500 }
 
-      context 'when data does not already exist for this period' do 
+      context 'when data does not already exist for this period' do
         let(:ts) { widget.decorate.series_end.advance(months: 1) }
         let(:schema) { slice_response_schema }
         include_examples 'api_authorized_ok'
@@ -46,20 +46,20 @@ RSpec.describe Api::V1::SlicesController, type: :controller do
         specify do
           [ds_1, ds_2].each do |dataset|
             row = widget.data_table.data_rows.by_time.last
-            expect(row.value_for(dataset)).to eq 500 
+            expect(row.value_for(dataset)).to eq 500
             expect(row.row_date).to eq ts
           end
         end
       end
 
-      context 'when data already exists for this period' do 
+      context 'when data already exists for this period' do
         let(:ts) { widget.decorate.series_start.to_date }
         specify { expect(response).to have_http_status 422 }
       end
     end
 
-    context 'invalid value' do 
-      context 'empty string' do 
+    context 'invalid value' do
+      context 'empty string' do
         let(:data_value) { '' }
         let(:ts) { widget.decorate.series_end.advance(months: 1) }
 
@@ -68,7 +68,7 @@ RSpec.describe Api::V1::SlicesController, type: :controller do
     end
   end
 
-  describe '#update' do 
+  describe '#update' do
     let(:ts) { widget.decorate.series_start.to_date }
     let(:req) { post :update, params: { widget_id: widget.id, period: 'month',
       period_start: ts,
@@ -83,7 +83,7 @@ RSpec.describe Api::V1::SlicesController, type: :controller do
     specify do
       [ds_1, ds_2].each do |dataset|
         row = widget.data_table.data_rows.by_time.first
-        expect(row.value_for(dataset)).to eq 500 
+        expect(row.value_for(dataset)).to eq 500
         expect(row.row_date).to eq ts
       end
     end
