@@ -1,4 +1,4 @@
-FactoryGirl.define do
+FactoryBot.define do
   factory :token do
     factory :token_expired do
       expired_at { rand(30).hours.ago }
