@@ -7,10 +7,10 @@ RSpec.describe DataTable, type: :model do
   it { is_expected.to have_many :datasets }
   it { is_expected.to have_many :data_rows }
 
-  describe 'metadata' do 
-    subject { FactoryGirl.create(:data_table, rows_count: 4) }
+  describe 'metadata' do
+    subject { FactoryBot.create(:data_table, rows_count: 4) }
 
-    describe 'series start and end' do 
+    describe 'series start and end' do
       its(:series_start) { is_expected.to eq subject.data_rows.by_time.first.row_date }
       its(:series_end) { is_expected.to eq subject.data_rows.last.row_date }
     end
@@ -21,8 +21,8 @@ RSpec.describe DataTable, type: :model do
     end
   end
 
-  describe '#slice_data' do 
-    let!(:widget) { FactoryGirl.create(:widget_with_data, datasets_count: 2, 
+  describe '#slice_data' do
+    let!(:widget) { FactoryBot.create(:widget_with_data, datasets_count: 2,
       rows_count: 4) }
     let(:data_table) { widget.data_table }
 
@@ -34,7 +34,7 @@ RSpec.describe DataTable, type: :model do
 
     let(:first_row) { data_table.data_rows.by_time.first }
 
-    context 'simple case, no aggregation' do 
+    context 'simple case, no aggregation' do
       let(:period) { 'month' }
 
       it { is_expected.to eq({
@@ -43,7 +43,7 @@ RSpec.describe DataTable, type: :model do
       }) }
     end
 
-    context 'aggregation' do 
+    context 'aggregation' do
       let(:period) { 'century' }
 
       def sum_value(data_table, dataset)
@@ -52,7 +52,7 @@ RSpec.describe DataTable, type: :model do
         end
       end
 
-      context 'mean' do 
+      context 'mean' do
         # Default slice aggregation is 'mean'
         it { is_expected.to eq({
           ds1.id => sum_value(data_table, ds1) / data_table.data_rows.count,
@@ -60,8 +60,8 @@ RSpec.describe DataTable, type: :model do
         }) }
       end
 
-      context 'sum' do 
-        before do 
+      context 'sum' do
+        before do
           data_table.slice_aggregation = 'sum'
         end
 
@@ -71,8 +71,8 @@ RSpec.describe DataTable, type: :model do
         }) }
       end
 
-      context 'with nil values' do 
-        before do 
+      context 'with nil values' do
+        before do
           data_table.data_rows.each do |row|
             data_table.datasets.each do |ds|
               row.set_value_for ds, nil
@@ -86,22 +86,22 @@ RSpec.describe DataTable, type: :model do
           ds1.id => nil,
           ds2.id => nil
         }) }
-      end 
+      end
     end
   end
 
-  describe '#slices' do 
-    let(:widget) { FactoryGirl.create :widget_with_data, datasets_count: 2,
+  describe '#slices' do
+    let(:widget) { FactoryBot.create :widget_with_data, datasets_count: 2,
       rows_count: 4 }
     let!(:data_table) { widget.data_table }
 
-    context 'Without limit' do 
+    context 'Without limit' do
       subject { data_table.slices widget }
       specify { expect(subject.size).to eq 4 }
       specify { expect(subject.first.period_start).to be < subject.last.period_start }
     end
 
-    context 'With limit' do 
+    context 'With limit' do
       subject { data_table.slices widget, limit: 2 }
       specify { expect(subject.size).to eq 2 }
       specify { expect(subject.first.period_start).to be < subject.last.period_start }
