@@ -4,7 +4,7 @@ module ControllerMacros
   def login_user
     before(:each) do
       @request.env["devise.mapping"] = Devise.mappings[:user]
-      user = FactoryGirl.create(:user_confirmed)
+      user = FactoryBot.create(:user_confirmed)
 
       sign_in user
     end
@@ -13,7 +13,7 @@ module ControllerMacros
   def logout_user
     before(:each) do
       @request.env["devise.mapping"] = Devise.mappings[:user]
-      user = FactoryGirl.create(:user_confirmed)
+      user = FactoryBot.create(:user_confirmed)
 
       sign_out user
     end
